change file names for better diff
