--- Amplify/Categories/DataStore/Model/Temporal/TimeZone+UTC.swift
@@ -1,15 +1,0 @@
-//
-// Copyright Amazon.com Inc. or its affiliates.
-// All Rights Reserved.
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-import Foundation
-
-extension TimeZone {
-    /// Utility UTC ("Coordinated Universal Time") TimeZone instance.
-    public static var utc: TimeZone {
-        TimeZone(abbreviation: "UTC")!
-    }
-}
